pyproject.toml
[project]
name = "qq-agent"
version = "0.1.0"
description = "QQ agent based on OneBot v11 + HelloAgents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "hello-agents>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qq_agent"]
